EPMDJ-7904 story: Redesign database architecture EPMDJ-8997 Integrate postgres DSM EPMDJ-9090 Refactoring and minor issues with new DB EPMDJ-9082 Changed embedded DB to test container
