+++ admin-part/build.gradle.kts
@@ -12,6 +12,8 @@ configurations.implementation {
     extendsFrom(jarDeps)
 }
 
+val drillDsmVersion: String by extra
+val testContainerVersion: String by project
 kotlin.sourceSets.all {
     languageSettings.useExperimentalAnnotation("kotlinx.coroutines.ExperimentalCoroutinesApi")
     languageSettings.useExperimentalAnnotation("kotlinx.serialization.ExperimentalSerializationApi")
@@ -37,6 +39,7 @@ dependencies {
 
     //provided by admin
     //TODO create a platform for admin dependencies
+    implementation("com.epam.drill:dsm:$drillDsmVersion")
     implementation("com.epam.drill:kodux")
     implementation("org.jetbrains.xodus:xodus-entity-store")
     implementation("org.jetbrains.kotlinx:kotlinx-collections-immutable")
@@ -44,6 +47,8 @@ dependencies {
     testImplementation(kotlin("test-junit5"))
     testImplementation("org.junit.jupiter:junit-jupiter")
     testImplementation("org.jetbrains.kotlinx:atomicfu")
+    testImplementation("org.slf4j:slf4j-simple:1.7.32")
+    testImplementation("org.testcontainers:postgresql:$testContainerVersion")
 }
 
 tasks {
