+++ admin-part/src/test/resources/logback.xml
@@ -1,0 +1,15 @@
+<configuration>
+    <appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
+        <encoder>
+            <pattern>%d{YYYY-MM-dd HH:mm:ss.SSS} [%thread] %-5level %logger{36} - %msg%n</pattern>
+        </encoder>
+    </appender>
+    <variable name="LOG_LEVEL" value="${LOG_LEVEL:-DEBUG}" />
+    <root level="${LOG_LEVEL}">
+        <appender-ref ref="STDOUT"/>
+    </root>
+    <logger name="org.eclipse.jetty" level="info"/>
+    <logger name="io.netty" level="info"/>
+    <logger name="io.mockk.impl.recording.states" level="off"/>
+    <logger name="jetbrains.exodus" level="off"/>
+</configuration>
